Create content for the Apple, ApplePicker, and AppleTree scripts.
Apple script is able to detect when it is to be destroyed when it touches the bottom of the playzone. ApplePicker script is connected to the camera and contains the management of the game by managing known apples and deleting them on failure, and managing the number of baskets and deleting them on failure. AppleTree script calls the methods DropApple every second after two seconds of starting and updates to move the tree every frame with a random chance of 2% every 1/50th second.